Aula03_Project_Bugflow/chat_interface/requirements.txt
streamlit>=1.28.0
pinecone-client>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
